corrections in master admin pages

+++ TheClinicApp1.1/Admin/Admin.aspx.cs
@@ -343,7 +343,10 @@ namespace TheClinicApp1._1.Admin
             {
                 //DeleteAssignedRoleByUserID(UserID);
 
-                DeleteUserByUserID(UserID);
+                if (hdnDeleteButtonClick.Value == "True")
+                {
+                    DeleteUserByUserID(UserID);
+                }
             }
             
         }
